requirements.txt
pydantic[email]==1.8.2
fastapi==0.68.0
uvicorn==0.14.0
SQLAlchemy==1.4.23
pydantic==1.8.2
sqlalchemy-utils==0.37.8
psycopg2-binary==2.9.1
python-dotenv==0.18.0
faker==8.10.1
fastapi-pagination==0.8.1
alembic==1.6.5
requests==2.22.0
pytest==6.2.4
elasticsearch[async]==7.14.0
databases==0.5.1
aioredis[hiredis]
asyncpg
httpx
pyjwt
